pyclient/requirements.txt
grpcio==1.67.0
protobuf==5.28.3
